Added Menu Page to Sample app that then call any of the two test pages.
