+++ PhotoViewerTest/MenuPage.xaml
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="PhotoViewerTest.MenuPage">
+  
+    <StackLayout Orientation="Vertical" VerticalOptions="Center" HorizontalOptions="CenterAndExpand" Spacing="0" Padding="10,10,10,10">
+      <Button Text="SingleImage" HorizontalOptions="FillAndExpand" VerticalOptions="Center" Clicked="OnButtonShowPageSingleImage" WidthRequest="200"/>
+      <Button Text="PhotoCarousel" HorizontalOptions="FillAndExpand" VerticalOptions="Center" Clicked="OnButtonShowPagePhotoCarousel" WidthRequest="200"/>
+    </StackLayout>
+</ContentPage>
