--- PhotoViewerTest/PhotoViewerTest.cs
@@ -1,32 +1,0 @@
-﻿using System;
-
-using Xamarin.Forms;
-
-namespace PhotoViewerTest
-{
-    public class App : Application
-    {
-        public App()
-        {
-            // The root page of your application
-            MainPage = new PhotoViewerTest.MainPage();
-            //MainPage = new PhotoViewerTest.SingleImagePage();
-        }
-
-        protected override void OnStart()
-        {
-            // Handle when your app starts
-        }
-
-        protected override void OnSleep()
-        {
-            // Handle when your app sleeps
-        }
-
-        protected override void OnResume()
-        {
-            // Handle when your app resumes
-        }
-    }
-}
-
